first revision of Persian and English parser libraries, they almost doesn't do anything right now!

+++ TranslatorBase/ttDictionary.cpp
@@ -53,7 +53,7 @@ bool ttDictionary::addWord(const ttArray& pWords,ttArray& dictionaries)
 		needNewSlot = true;
 	else
 		for(unsigned i=0;i<pWords.objNum() && needNewSlot == false;i++)
-			if (wID != i)
+			if (wID != (int)i)
 				if (static_cast<ttDictionary*>(dictionaries[i])->mWordList->operator[](wID) != NULL)
 					//if (*static_cast<ttDictionary*>(dictionaries[i])->mWordList->operator[](wID) != (*static_cast<ttWord*>(pWords[i])))
 						needNewSlot = true;
